backend2/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
yfinance==0.2.51
pandas==2.2.3
numpy==1.26.4
scipy==1.13.1
google-cloud-firestore==2.19.0
pydantic==2.9.2
